requirements.txt
fastapi
langgraph
faiss-cpu
yfinance
langchain-google-genai
sortedcontainers